@@ -4,67 +4,66 @@ description: An application can help you collect game data more easily via
 
 publish_to: none
 
-version: 0.1.2+1
+version: 0.1.3+1
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: ">=3.0.6 <4.0.0"
 
 dependencies:
   another_flushbar: 1.12.30
   badges: 3.1.1
-  bot_toast: 4.0.3
+  bot_toast: 4.0.4
   collection: 1.17.1
-  connectivity_plus: 4.0.0
-  dartx: 1.1.0
+  connectivity_plus: 4.0.1
+  dartx: 1.2.0
   dartz: 0.10.1
-  data_table_2: 2.5.0
-  dio: 5.1.2
-  drift: 2.8.0
-  easy_localization: 3.0.1
+  data_table_2: 2.5.6
+  dio: 5.3.1
+  drift: 2.10.0
+  easy_localization: 3.0.2
   ffi: 2.0.2
-  file_picker: 5.3.0
-  fluent_ui: 4.6.1
+  file_picker: 5.3.3
+  fluent_ui: 4.7.1
   flutter:
     sdk: flutter
   flutter_riverpod: 2.3.6
-  flutter_screenutil: 5.8.1
-  font_awesome_flutter: 10.4.0
-  freezed_annotation: 2.2.0
-  go_router: 7.0.1
-  intl: 0.17.0
+  flutter_screenutil: 5.8.4
+  font_awesome_flutter: 10.5.0
+  freezed_annotation: 2.4.1
+  go_router: 10.0.0
+  intl: 0.18.0
   json_annotation: 4.8.1
-  logger: 1.3.0
+  logger: 2.0.1
   modal_bottom_sheet: 3.0.0-pre
-  package_info_plus: 4.0.0
+  package_info_plus: 4.0.2
   path: 1.8.3
   path_provider: 2.0.15
   pretty_dio_logger: 1.3.1
   retrofit: 4.0.1
-  shared_preferences: 2.1.1
+  riverpod_annotation: 2.1.1
+  shared_preferences: 2.2.0
   sqlite3_flutter_libs: 0.5.15
   time: 2.1.3
-  url_launcher: 6.1.11
-  webview_windows: 0.2.1
-  win32: 4.1.4
-  window_manager: 0.3.2
+  url_launcher: 6.1.12
+  webview_windows: 0.3.0
+  win32: 5.0.6
+  window_manager: 0.3.5
 
 dev_dependencies:
-  build_runner: 2.4.4
+  build_runner: 2.4.6
   build_verify: 3.1.0
   custom_lint: 0.4.0
-  drift_dev: 2.8.1
+  drift_dev: 2.10.0
   flutter_gen_runner: 5.3.1
   flutter_test:
     sdk: flutter
-  freezed: 2.3.4
-  go_router_builder: 2.0.1
-  json_serializable: 6.6.2
+  freezed: 2.4.1
+  go_router_builder: 2.2.4
+  json_serializable: 6.7.1
   lint: 2.1.2
-  retrofit_generator: 6.0.0+3
-  riverpod_lint: 1.3.2
-
-dependency_overrides:
-  intl: 0.18.1
+  retrofit_generator: 7.0.2
+  riverpod_generator: 2.2.4
+  riverpod_lint: 1.4.0
 
 flutter_gen:
   output: lib/core/gen
