@@ -1,4 +1,7 @@
+import 'package:easy_localization/easy_localization.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
+
+import '../../generated/locale_keys.g.dart';
 
 part 'app_failure.freezed.dart';
 
@@ -19,21 +22,23 @@ class AppFailure with _$AppFailure {
 
   const factory AppFailure.localDataError(Object e) = _LocalDataError;
 
-  const factory AppFailure.localDataEmpty() = _LocalDataEmpty;
+  const factory AppFailure.emptyLocalData() = _EmptyLocalData;
 
   const factory AppFailure.invalidToken() = _InvalidToken;
 
   String get localizedMessage => map(
         localizedError: (f) => f.message,
-        unexpectedError: (f) => '未知错误\n ${f.e}',
-        networkUnreachable: (_) => '系统检测到网络异常，请确保您的网线没有被猫咬断。',
+        unexpectedError: (f) =>
+            '${LocaleKeys.error_unexpectedError.tr()}\n${f.e}',
+        networkUnreachable: (_) => LocaleKeys.error_networkUnreachable.tr(),
         remoteServerError: (f) {
           final status = f.code == null ? '' : '[${f.code}]';
-          final message = f.message ?? '服务器软体不稳定，请稍后重试。';
+          final message = f.message ?? LocaleKeys.error_remoteServerError.tr();
           return '$status$message';
         },
-        localDataError: (f) => '无法加载本地数据，请稍后重试。\n${f.e}',
-        localDataEmpty: (_) => '本地数据不存在，请尝试重新获取数据。',
-        invalidToken: (_) => '令牌错误或已失效，请重新登录获取最新的访问令牌。',
+        localDataError: (f) =>
+            '${LocaleKeys.error_localDataError.tr()}\n${f.e}',
+        emptyLocalData: (_) => LocaleKeys.error_emptyLocalData.tr(),
+        invalidToken: (_) => LocaleKeys.error_invalidToken.tr(),
       );
 }
