@@ -1,8 +1,10 @@
+import 'package:easy_localization/easy_localization.dart';
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 import '../../../application/gacha/gacha_pool_selector_provider.dart';
+import '../../../generated/locale_keys.g.dart';
 
 class GachaPoolSelector extends ConsumerWidget {
   const GachaPoolSelector({super.key});
@@ -14,7 +16,7 @@ class GachaPoolSelector extends ConsumerWidget {
         .map(
           (pool) => MenuFlyoutItem(
             text: Text(
-              pool ?? '全部寻访',
+              pool ?? LocaleKeys.gacha_type_all.tr(),
               style: DefaultTextStyle.of(context).style,
             ),
             onPressed: () => provider.select(pool),
@@ -24,7 +26,7 @@ class GachaPoolSelector extends ConsumerWidget {
     return DropDownButton(
       items: items,
       title: Text(
-        provider.selectedPool ?? '全部寻访',
+        provider.selectedPool ?? LocaleKeys.gacha_type_all.tr(),
         style: TextStyle(fontSize: 16.sp),
       ),
     );
